+++ controller/mydict.js
@@ -1,5 +1,6 @@
 const Mydict = require("../models/mydict");
 const Opendict = require("../models/opendict");
+const { findMydictMeanings } = require("../lib/mydict/findMydictMeanings");
 
 // 나만의 단어장 등록하기
 const postMydict = async (req, res) => {
@@ -79,7 +80,7 @@ const getMydictSome = async (req, res) => {
       return res.json({ ok: false, errorMessage: "등록한 단어가 없습니다." });
     }
 
-    let mydictMeanings = await getMydictMeanings(req, res);
+    let mydictMeanings = await findMydictMeanings(req, res);
     mydictMeanings = mydictMeanings.reverse().slice(0, 4);
 
     res.json({
@@ -109,7 +110,7 @@ const getMydictAll = async (req, res) => {
       return res.json({ ok: false, errorMessage: "등록한 단어가 없습니다." });
     }
 
-    let mydictMeanings = await getMydictMeanings(req, res);
+    let mydictMeanings = await findMydictMeanings(req, res);
     mydictMeanings = mydictMeanings.reverse();
 
     res.json({
@@ -158,152 +159,6 @@ const deleteMydict = async (req, res) => {
     console.error(`${error} 에러로 나만의 단어장 단어 삭제 실패`);
   }
 };
-
-// function : 나만의 단어장 단어(단어 뜻)가져오기
-async function getMydictMeanings(req, res) {
-  const nickname = res.locals.user.nickname;
-
-  // 나만의 단어장에서 등록한 단어 찾기
-  const findMydictWords = await Mydict.aggregate([
-    { $match: { nickname } },
-    { $project: { _id: 0, nickname: 1, scriptId: 1, word: 1, sentence: 1 } },
-  ]);
-
-  // 나만의 단어장에 등록된 단어로 오픈사전 단어장에서 좋아요 1위 단어 찾기
-  const findMostLikedMeaning = await Mydict.aggregate([
-    { $match: { nickname } },
-    {
-      $lookup: {
-        from: "opendicts",
-        localField: "word",
-        foreignField: "word",
-        let: { scriptId: "$scriptId" },
-        pipeline: [
-          {
-            $match: {
-              $expr: { $in: ["$scriptId", ["$$scriptId"]] },
-            },
-          },
-          {
-            $project: {
-              _id: 0,
-              word: 1,
-              meaning: 1,
-              count: 1,
-              nickname: 1,
-              scriptId: 1,
-            },
-          },
-          {
-            $sort: { count: -1 },
-          },
-          {
-            $group: {
-              _id: "$word",
-              meaning: { $first: "$meaning" },
-            },
-          },
-        ],
-        as: "meaning",
-      },
-    },
-    {
-      $project: {
-        _id: 0,
-        meaning: 1,
-      },
-    },
-  ]);
-
-  // 나만의 단어장에 등록된 단어로 오픈사전 단어장에서 내가 등록한 단어 찾기
-  const findMydictWord = await Mydict.aggregate([
-    { $match: { nickname } },
-    {
-      $lookup: {
-        from: "opendicts",
-        localField: "word",
-        foreignField: "word",
-        let: { scriptId: "$scriptId", nickname: "$nickname" },
-        pipeline: [
-          {
-            $match: {
-              $expr: {
-                $and: [
-                  { $in: ["$scriptId", ["$$scriptId"]] },
-                  { $in: ["$nickname", ["$$nickname"]] },
-                ],
-              },
-            },
-          },
-          {
-            $project: {
-              _id: 0,
-              word: 1,
-              meaning: 1,
-              count: 1,
-              nickname: 1,
-              scriptId: 1,
-            },
-          },
-          {
-            $sort: { count: -1 },
-          },
-          {
-            $group: {
-              _id: "$word",
-              meaning: { $first: "$meaning" },
-            },
-          },
-        ],
-        as: "meaning",
-      },
-    },
-    {
-      $project: {
-        _id: 0,
-        meaning: 1,
-      },
-    },
-  ]);
-
-  // 나만의 단어장에 찾은 단어 넣는 곳
-  let mydictMeanings = [];
-
-  // 나만의 단어장에 좋아요 1위 단어 넣기
-  for (let mydictMeaning of findMostLikedMeaning) {
-    let mostLikedMeaning = mydictMeaning.meaning;
-    mydictMeanings.push(mostLikedMeaning);
-  }
-
-  // 나만의 단어장에 내가 저장한 단어 넣기
-  let idx = 0;
-  for (let mydictMeaning2 of findMydictWord) {
-    let myMeaning = mydictMeaning2.meaning;
-
-    // 프론트로 정제해서 보내기
-    // 내가 등록한 단어 뜻이 없을 때 빈 칸으로 넣어줌
-    if (myMeaning.length === 0) {
-      myMeaning = [{ meaning: "" }];
-    }
-
-    // 좋아요 1위 단어 뜻 옆에 넣기
-    mydictMeanings[idx].push(
-      ...myMeaning,
-      findMydictWords[idx].word,
-      findMydictWords[idx].sentence,
-      findMydictWords[idx].scriptId
-    );
-
-    // 객체 분리해서 넣어주기
-    if (mydictMeanings[idx][0] || mydictMeanings[idx][1]) {
-      mydictMeanings[idx][0] = mydictMeanings[idx][0].meaning;
-      mydictMeanings[idx][1] = mydictMeanings[idx][1].meaning;
-    }
-
-    idx++;
-  }
-  return mydictMeanings;
-}
 
 module.exports = {
   postMydict,
